fix: all comments

+++ Client/ConsoleCommands.cs
@@ -1,5 +1,8 @@
 using System;
 using ToDoList.Services;
+using ToDoList.Tools;
+using static System.DateTime;
+
 namespace ToDoList.Client
 {
     public class ConsoleCommands
@@ -8,14 +11,14 @@ namespace ToDoList.Client
         private readonly Drawing _drawing;
         private bool _run;
         private string[] _input;
-        public ConsoleCommands()
+        public ConsoleCommands(IToDoListService service, Drawing drawing)
         {
-            _service = new ToDoListService();
-            _drawing = new Drawing();
+            _service = service;
+            _drawing = drawing;
             _run = true;
         }
 
-        private void ReadCommands()
+        private void ReadAndExecuteCommands()
         {
             _input = Console.ReadLine().Split(" ");
             var args = string.Join(' ', _input, 1, _input.Length - 1);
@@ -29,26 +32,31 @@ namespace ToDoList.Client
                     _drawing.DrawTable(_service.GetTasksNotInGroup());
                     break;
                 case "/delete":
+                    if (_input.Length != 2) throw new ToDoListException("args length should be 1");
                     _service.DeleteTaskById(args);
                     break;
                 case "/save":
-                    _service.Save();
+                    _service.Save(args);
                     break;
                 case "/load":
-                    _service.Load();
+                    _service.Load(args);
                     _drawing.DrawGroupsTree(_service.GetAllGroups());
                     _drawing.DrawTable(_service.GetTasksNotInGroup());
                     break;
                 case "/complete":
+                    if (_input.Length != 2) throw new ToDoListException("args length should be 1");
                     _service.DoComplete(args);
                     break;
                 case "/completed":
                     _drawing.DrawTable(_service.GetAllCompletedTasks());
                     break;
                 case "/deadline":
+                    if (_input.Length != 3) throw new ToDoListException("args length should be 2");
                     var deadlineArgs = args.Split(" ");
                     var deadLineTime = string.Join(' ', deadlineArgs, 1, deadlineArgs.Length - 1);
-                    _service.AddDeadLine(deadlineArgs[0], DateTime.Parse(deadLineTime));
+                    if (!TryParse(deadLineTime, out var deadline))
+                        throw new ToDoListException("cant parse argument");
+                    _service.AddDeadLine(deadlineArgs[0], deadline);
                     break;
                 case "/today":
                     _drawing.DrawTable(_service.GetTodayTasks());
@@ -62,14 +70,17 @@ namespace ToDoList.Client
                     _service.AddGroup(args);
                     break;
                 case "/delete-group":
+                    if (_input.Length != 2) throw new ToDoListException("args length should be 1");
                     _service.DeleteGroup(args);
                     break;
                 case "/add-to-group":
+                    if (_input.Length != 3) throw new ToDoListException("args length should be 2");
                     var groupArgs = args.Split(" ");
                     var groupName = string.Join(' ', groupArgs, 1, groupArgs.Length - 1);
                     _service.AddTaskToGroup(groupArgs[0], groupName);
                     break;
                 case "/delete-from-group":
+                    if (_input.Length != 3) throw new ToDoListException("args length should be 2");
                     var delGroupArgs = args.Split(" ");
                     var delGroupName = string.Join(' ', delGroupArgs, 1, delGroupArgs.Length - 1);
                     _service.DeleteTaskFromGroup(delGroupArgs[0], delGroupName);
@@ -87,7 +98,7 @@ namespace ToDoList.Client
         {
             while (_run)
             {
-                ReadCommands();
+                ReadAndExecuteCommands();
             }
         }
     }
